fix: move hook to separate file
